Fixed some outstanding issues with drag and drop:
1. The Safari quirks code for drag and drop was using editor.insertContent.
This bypassed the PastePreProcess and PastePostProcess events. Switched
this to use the mceInsertClipboardContent command instead, which goes
through the paste plugin and fires the correct events.

2. IE does not support setting custom contentType in the dataTransfer
object, which meant drag and drop between editors did not work at all on
IE. I extended the existing use of the special data:text/mce-internal url
to IE so that it can now work.

3. Fixed a couple issues in the paste plugin:
  - BMP images are now allowed.
  - Multiple images can be dropped in one drag.